refactor: optimize Music Stack Interaction widget styles and structure
